requirements.txt
amqp==2.3.2
anyjson==0.3.3
beautifulsoup4==4.6.3
billiard==3.5.0.4
celery==4.2.1
confusable-homoglyphs==3.2.0
Django>1.11.0,<1.12
django-appconf==1.0.2
django-axes==4.4.2
django-cors-headers==2.4.0
django-ipware==2.1.0
django-ranged-response==0.2.0
django-registration==3
django-reversion==2.0.13
django-reversion-compare==0.8.5
django-simple-captcha==0.5.9
djangorestframework==3.9.0
gitdb2==2.0.5
GitPython==2.1.11
kombu==4.2.1
mysqlclient==1.3.13
Pillow==5.3.0
pytz==2018.7
six==1.11.0
smmap2==2.0.5
vine==1.1.4
